Do not augument provided executor command.
This patch changes behavior with custom executor.
The executor command used to be wrapped with `chmod` to change permission
to launch it. This is not necessary, what's more this
could fail with Docker or system configuration where
executor permissions couldn't be changed by Mesos Agent.
This patch also removes command that used to be passed as
a argument to executor. There is no need for this since
all information about command to run will be delivered
to executor in LaunchTask event by Mesos Agent.

Fixes: [MARATHON-4210](https://jira.mesosphere.com/browse/MARATHON-4210)
Refs: http://mesos.apache.org/documentation/latest/executor-http-api/